+++ classes/item.rb
