+++ json/albums.json
@@ -1,27 +1,27 @@
 [
   {
-    "id": 851,
+    "id": 681,
     "genre": "Horror",
     "publish_date": "2020",
     "archived": true,
     "on_spotify": false
   },
   {
-    "id": 56,
+    "id": 606,
     "genre": "Thriller",
     "publish_date": "2012",
     "archived": true,
     "on_spotify": true
   },
   {
-    "id": 513,
+    "id": 219,
     "genre": "Action",
     "publish_date": "2023",
     "archived": false,
     "on_spotify": false
   },
   {
-    "id": 95,
+    "id": 944,
     "genre": "Romance",
     "publish_date": "207",
     "archived": true,
